% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_continue.R
\name{ask_continue}
\alias{ask_continue}
\title{Ask Continue}
\usage{
ask_continue(msg = "Press Enter to continue: ", begin = "\\n",
  end = "\\n")
}
\arguments{
\item{msg}{A message to display.}
}
\description{
Propmt for the user to press Enter to continue.
}
\examples{
ask_continue()
}
\keyword{continue}
