% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cmd_cwd.R
\name{get_cmd_cwd}
\alias{get_cmd_cwd}
\title{Get Command's CWD}
\usage{
get_cmd_cwd()
}
\description{
Get the command's cwd for local file access.
}
\examples{
get_cmd_cwd()
}
\keyword{directory}
\keyword{working}
