% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inform_user.R
\name{inform_about}
\alias{inform_about}
\title{Inform User}
\usage{
inform_about(title = "", text = "", delim = "=", begin = "",
  end = "\\n")
}
\arguments{
\item{title}{The section title.}

\item{text}{The text to display.}

\item{delim}{Delimiter to highlight the title.}

\item{begin}{Text at the beginning.}

\item{end}{Text at the end.}
}
\description{
Display an informative message.
}
\examples{
inform_user()
}
\keyword{informative}
