% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlask.R
\name{mlask}
\alias{mlask}
\title{Ask Continue}
\usage{
mlask(msg = "Press Enter to continue: ", begin = "\\n", end = "\\n")
}
\arguments{
\item{msg}{A message to display.}
}
\description{
Prompt for the user to press Enter to continue.
}
\examples{
mlask()
}
\keyword{continue}
