% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcat.R
\name{mlcat}
\alias{mlcat}
\title{Inform User}
\usage{
mlcat(title = "", text = "", delim = "=", begin = "", end = "\\n")
}
\arguments{
\item{title}{The section title.}

\item{text}{The text to display after processing with glue.}

\item{delim}{Delimiter to highlight the title.}

\item{begin}{Text at the beginning.}

\item{end}{Text at the end.}
}
\description{
Display an informative message.
}
\examples{
mlcat()
}
\keyword{informative}
