% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpreview.R
\name{mlpreview}
\alias{mlpreview}
\title{Preview File}
\usage{
mlpreview(
  fname,
  msg = "Close the graphic window using Ctrl-W.\\n",
  begin = "\\n",
  previewer = "xdg-open"
)
}
\arguments{
\item{fname}{File name to display}

\item{msg}{Message to display after displaying the file.}

\item{previewer}{The command to run to display the file.}
}
\description{
Display PDF file.
}
\examples{
mlpreview()
}
\keyword{display}
