% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_file.R
\name{preview_file}
\alias{preview_file}
\title{Preview File}
\usage{
preview_file(fname, previewer = "atril --preview")
}
\arguments{
\item{fname}{File name to display}
}
\description{
Display PDF file.
}
\examples{
show_close_graphics_msg()
}
\keyword{display}
